+++ src/endpoints/category/service/service-category.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, UseInterceptors } from '@nestjs/common';
-import { SentryInterceptor } from 'src/common';
+import { SentryInterceptor } from '../../../common';
 import { ServiceCategoryService } from './service-category.service';
 
 @UseInterceptors(SentryInterceptor)
