+++ src/endpoints/recaptcha/recaptcha.controller.ts
@@ -1,7 +1,7 @@
 import { Body, Controller, Post, Res, UseInterceptors } from '@nestjs/common';
 import axios from 'axios';
 import { Response } from 'express';
-import { SentryInterceptor } from 'src/common';
+import { SentryInterceptor } from '../../common';
 
 @UseInterceptors(SentryInterceptor)
 @Controller('recaptcha')
