+++ src/endpoints/substrate-indexed-data/labs/lab.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, Query, UseInterceptors } from '@nestjs/common';
 import { ApiQuery } from '@nestjs/swagger';
-import { SentryInterceptor } from 'src/common';
+import { SentryInterceptor } from '../../../common';
 import { LabService } from './lab.service';
 
 @UseInterceptors(SentryInterceptor)
