+++ src/endpoints/verification/verification.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { DateTimeModule } from 'src/common/date-time/date-time.module';
-import { ProcessEnvModule } from 'src/common/process-env/process-env.module';
-import { RewardModule } from 'src/reward/reward.module';
-import { SubstrateModule } from 'src/substrate/substrate.module';
+import { DateTimeModule } from '../../common/date-time/date-time.module';
+import { ProcessEnvModule } from '../../common/process-env/process-env.module';
+import { RewardModule } from '../../reward/reward.module';
+import { SubstrateModule } from '../../substrate/substrate.module';
 import { VerificationController } from './verification.controller';
 import { VerificationService } from './verification.service';
