+++ src/substrate/orderEvent.ts
@@ -1,6 +1,6 @@
 import { ApiPromise } from '@polkadot/api';
 import { Option } from '@polkadot/types';
-import { EscrowService } from 'src/endpoints/escrow/escrow.service';
+import { EscrowService } from '../endpoints/escrow/escrow.service';
 import { RewardDto } from 'src/reward/dto/reward.dto';
 import { RewardService } from 'src/reward/reward.service';
 import { SubstrateService } from './substrate.service';
@@ -8,7 +8,7 @@ import { Logger } from '@nestjs/common';
 import { TransactionLoggingService } from 'src/transaction-logging/transaction-logging.service';
 import { TransactionLoggingDto } from 'src/transaction-logging/dto/transaction-logging.dto';
 import { ethers } from 'ethers'
-import { DebioConversionService } from 'src/debio-conversion/debio-conversion.service';
+import { DebioConversionService } from '../debio-conversion/debio-conversion.service';
 
 export class OrderEventHandler {
   constructor(
