+++ src/substrate/substrate.module.ts
@@ -3,10 +3,10 @@ import { SubstrateService } from './substrate.service';
 import { SubstrateController } from './substrate.controller';
 import { EscrowModule } from '../endpoints/escrow/escrow.module';
 import { TransactionLoggingModule } from '../transaction-logging/transaction-logging.module';
-import { RewardModule } from 'src/reward/reward.module';
-import { MailModule } from 'src/common/mailer/mailer.module';
-import { LocationModule } from 'src/endpoints/location/location.module';
-import { DebioConversionModule } from 'src/debio-conversion/debio-conversion.module';
+import { RewardModule } from '../reward/reward.module';
+import { MailModule } from '../common/mailer/mailer.module';
+import { LocationModule } from '../endpoints/location/location.module';
+import { DebioConversionModule } from '../debio-conversion/debio-conversion.module';
 
 @Module({
   imports: [
