+++ test/unit/endpoints/category/emr/emr.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { EmrController } from '../../../../src/endpoints/category/emr/emr.controller';
-import { EmrService } from '../../../../src/endpoints/category/emr/emr.service';
-import { MockType } from '../../mock';
+import { EmrController } from '../../../../../src/endpoints/category/emr/emr.controller';
+import { EmrService } from '../../../../../src/endpoints/category/emr/emr.service';
+import { MockType } from '../../../mock';
 
 describe('Emr Controller Unit Tests', () => {
   const emrServiceMockFactory: () => MockType<EmrService> = jest.fn(() => ({
