+++ test/unit/endpoints/category/emr/emr.service.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
-import { EmrService } from '../../../../src/endpoints/category/emr/emr.service';
-import { EmrCategory } from '../../../../src/endpoints/category/emr/models/emr.entity';
-import { MockType, repositoryMockFactory } from '../../mock';
+import { EmrService } from '../../../../../src/endpoints/category/emr/emr.service';
+import { EmrCategory } from '../../../../../src/endpoints/category/emr/models/emr.entity';
+import { MockType, repositoryMockFactory } from '../../../mock';
 import { Repository } from 'typeorm';
 
 describe('Emr Service Unit Tests', () => {
