+++ test/unit/endpoints/location/state.service.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
-import { StateService } from '../../../src/endpoints/location/state.service';
-import { State } from '../../../src/endpoints/location/models/state.entity';
-import { MockType, repositoryMockFactory } from '../mock';
+import { StateService } from '../../../../src/endpoints/location/state.service';
+import { State } from '../../../../src/endpoints/location/models/state.entity';
+import { MockType, repositoryMockFactory } from '../../mock';
 import { Repository } from 'typeorm';
 import { when } from 'jest-when';
