+++ test/unit/endpoints/substrate-indexed-data/labs/lab.controller.spec.ts
@@ -1,10 +1,12 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { when } from 'jest-when';
 import { ElasticsearchService } from '@nestjs/elasticsearch';
-import { LabService } from '../../../../src/endpoints/substrate-indexed-data/labs/lab.service';
-import { elasticsearchServiceMockFactory, MockType } from '../../mock';
+import { LabService } from '../../../../../src/endpoints/substrate-indexed-data/labs/lab.service';
+import { LabController } from '../../../../../src/endpoints/substrate-indexed-data/labs/lab.controller';
+import { elasticsearchServiceMockFactory, MockType } from '../../../mock';
 
-describe('Substrate Indexer Lab Service Unit Tests', () => {
+describe('Substrate Indexer Lab Controller Unit Tests', () => {
+  let labControllerMock: LabController;
   let labServiceMock: LabService;
   let elasticsearchServiceMock: MockType<ElasticsearchService>;
 
@@ -51,22 +53,25 @@ describe('Substrate Indexer Lab Service Unit Tests', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
+        LabController,
         LabService,
         { provide: ElasticsearchService, useFactory: elasticsearchServiceMockFactory }
       ],
     }).compile();
 
+    labControllerMock = module.get(LabController);
     labServiceMock = module.get(LabService);
     elasticsearchServiceMock = module.get(ElasticsearchService);
   });
 
   it('should be defined', () => {
     // Assert
-    expect(labServiceMock).toBeDefined();
+    expect(labControllerMock).toBeDefined();
   });
 
   it('should find lab by country, city, and category', () => {
     // Arrange
+    const labServiceSpy = jest.spyOn(labServiceMock, 'getByCountryCityCategory');
     const CALLED_WITH = createSearchObject(
         "XX",
         "XX",
@@ -104,7 +109,7 @@ describe('Substrate Indexer Lab Service Unit Tests', () => {
       }
     };
     when(elasticsearchServiceMock.search).calledWith(CALLED_WITH).mockReturnValue(ES_RESULT);
-
+    
     const RESULT = {
       result: [
         {
@@ -124,7 +129,7 @@ describe('Substrate Indexer Lab Service Unit Tests', () => {
     }
 
     // Assert
-    expect(labServiceMock.getByCountryCityCategory(
+    expect(labControllerMock.findByCountryCityCategory(
         "XX",
         "XX",
         "XX",
@@ -133,6 +138,7 @@ describe('Substrate Indexer Lab Service Unit Tests', () => {
         1,
         10
     )).resolves.toEqual(RESULT);
+    expect(labServiceSpy).toHaveBeenCalled();
     expect(elasticsearchServiceMock.search).toHaveBeenCalled();
   });
 
@@ -146,20 +152,22 @@ describe('Substrate Indexer Lab Service Unit Tests', () => {
         }
       }
     };
+    const labServiceSpy = jest.spyOn(labServiceMock, 'getByCountryCityCategory');
     elasticsearchServiceMock.search.mockImplementationOnce(() => Promise.reject(ERROR_RESULT));
 
     // Assert
-    expect(labServiceMock.getByCountryCityCategory(
-        "XX",
-        "XX",
-        "XX",
-        "XX",
-        false,
-        1,
-        10
+    expect(labControllerMock.findByCountryCityCategory(
+      "XX",
+      "XX",
+      "XX",
+      "XX",
+      false,
+      1,
+      10
     )).resolves.toEqual({
       result: RESULT
     });
+    expect(labServiceSpy).toHaveBeenCalled();
     expect(elasticsearchServiceMock.search).toHaveBeenCalled();
   });
 
@@ -172,10 +180,11 @@ describe('Substrate Indexer Lab Service Unit Tests', () => {
         }
       }
     };
-    elasticsearchServiceMock.search.mockImplementationOnce(() => Promise.reject(ERROR_RESULT));
+    const labServiceSpy = jest.spyOn(labServiceMock, 'getByCountryCityCategory');
+    elasticsearchServiceMock.search.mockImplementationOnce(() => Promise.reject(ERROR_RESULT))
 
     // Assert
-    expect(labServiceMock.getByCountryCityCategory(
+    expect(labControllerMock.findByCountryCityCategory(
       "XX",
       "XX",
       "XX",
@@ -184,6 +193,7 @@ describe('Substrate Indexer Lab Service Unit Tests', () => {
       1,
       10
     )).rejects.toMatchObject(ERROR_RESULT);
+    expect(labServiceSpy).toHaveBeenCalled();
     expect(elasticsearchServiceMock.search).toHaveBeenCalled();
   });
 });
